+++ src/helpers/index.ts
@@ -41,3 +41,4 @@ export * from './trusted-types-utils';
 export * from './value-matchers';
 export * from './click-utils';
 export * from './set-constant-utils';
+export * from './chain-prop-utils';