+++ src/scriptlets/abort-on-property-read.js
@@ -5,6 +5,7 @@ import {
     createOnErrorHandler,
     hit,
     isEmptyObject,
+    interceptChainProp,
 } from '../helpers';
 
 /* eslint-disable max-len */
@@ -53,19 +54,9 @@ export function abortOnPropertyRead(source, property) {
     };
     const setChainPropAccess = (owner, property) => {
         const chainInfo = getPropertyInChain(owner, property);
-        let { base } = chainInfo;
-        const { prop, chain } = chainInfo;
+        const { base, prop, chain } = chainInfo;
         if (chain) {
-            const setter = (a) => {
-                base = a;
-                if (a instanceof Object) {
-                    setChainPropAccess(a, chain);
-                }
-            };
-            Object.defineProperty(owner, prop, {
-                get: () => base,
-                set: setter,
-            });
+            interceptChainProp(owner, prop, chain, setChainPropAccess);
             return;
         }
 
@@ -103,4 +94,5 @@ abortOnPropertyRead.injections = [
     createOnErrorHandler,
     hit,
     isEmptyObject,
+    interceptChainProp,
 ];