+++ src/scriptlets/abort-on-stack-trace.js
@@ -15,6 +15,7 @@ import {
     getNativeRegexpTest,
     backupRegExpValues,
     restoreRegExpValues,
+    interceptChainProp,
 } from '../helpers';
 
 /* eslint-disable max-len */
@@ -89,19 +90,9 @@ export function abortOnStackTrace(source, property, stack) {
 
     const setChainPropAccess = (owner, property) => {
         const chainInfo = getPropertyInChain(owner, property);
-        let { base } = chainInfo;
-        const { prop, chain } = chainInfo;
+        const { base, prop, chain } = chainInfo;
         if (chain) {
-            const setter = (a) => {
-                base = a;
-                if (a instanceof Object) {
-                    setChainPropAccess(a, chain);
-                }
-            };
-            Object.defineProperty(owner, prop, {
-                get: () => base,
-                set: setter,
-            });
+            interceptChainProp(owner, prop, chain, setChainPropAccess);
             return;
         }
 
@@ -177,4 +168,5 @@ abortOnStackTrace.injections = [
     shouldAbortInlineOrInjectedScript,
     backupRegExpValues,
     restoreRegExpValues,
+    interceptChainProp,
 ];