+++ tests/scriptlets/abort-current-inline-script.test.js
@@ -282,3 +282,13 @@ test('searches script by regexp - abort few inline scripts', (assert) => {
     assert.strictEqual(window.shouldNotBeAborted, true, 'value of shouldBeAborted has been changed from false to true');
     assert.strictEqual(window.hit, 'FIRED', 'hit fired');
 });
+
+test('check if chained property is undefined', (assert) => {
+    const property = 'foo.bar';
+    const search = 'test_qwerty';
+    const scriptletArgs = [property, search];
+    runScriptlet(name, scriptletArgs);
+
+    assert.strictEqual(window.foo, undefined, 'window.foo is undefined');
+    assert.strictEqual(window.hit, undefined, 'hit should NOT fire');
+});