+++ tests/scriptlets/debug-on-property-read.test.js
@@ -60,3 +60,12 @@ test('works with an empty object in chain', (assert) => {
     console.log(window.aaa.bbb);
     assert.strictEqual(window.hit, 'FIRED', 'hit fired');
 });
+
+test('check if chained property is undefined', (assert) => {
+    const property = 'foo.bar';
+    const scriptletArgs = [property];
+    runScriptlet(name, scriptletArgs);
+
+    assert.strictEqual(window.foo, undefined, 'window.foo is undefined');
+    assert.strictEqual(window.hit, undefined, 'hit should NOT fire');
+});