+++ tests/scriptlets/log-on-stack-trace.test.js
@@ -23,6 +23,15 @@ const afterEach = () => {
 
 module(name, { beforeEach, afterEach });
 
+test('check if chained property is undefined', (assert) => {
+    const property = 'foo.bar';
+    const scriptletArgs = [property];
+    runScriptlet(name, scriptletArgs);
+
+    assert.strictEqual(window.foo, undefined, 'window.foo is undefined');
+    assert.strictEqual(window.hit, undefined, 'hit should NOT fire');
+});
+
 test('can get property', (assert) => {
     window[PROPERTY] = 'value';
     const scriptletArgs = [PROPERTY];