+++ CHANGELOG.md
@@ -10,6 +10,25 @@ The format is based on [Keep a Changelog], and this project adheres to [Semantic
 <!-- TODO: change `@added unknown` tag due to the actual version -->
 <!--       during new scriptlets or redirects releasing -->
 
+## [Unreleased]
+
+### Changed
+
+- Added `interceptChainProp` helper to share intermediate chain property access logic
+  across `abort-on-property-read`, `abort-on-property-write`, `abort-on-stack-trace`,
+  `abort-current-inline-script`, `debug-on-property-write`, `debug-on-property-read`,
+  `debug-current-inline-script` and `log-on-stack-trace` scriptlets [#513].
+
+### Fixed
+
+- issue with assigning `window` to property which was not set in
+  `abort-on-property-read`, `abort-on-property-write`, `abort-on-stack-trace`,
+  `abort-current-inline-script`, `debug-on-property-write`, `debug-on-property-read`,
+  `debug-current-inline-script` and `log-on-stack-trace` scriptlets [#513].
+
+[Unreleased]: https://github.com/AdguardTeam/Scriptlets/compare/v2.2.16...HEAD
+[#513]: https://github.com/AdguardTeam/Scriptlets/issues/513
+
 ## [v2.2.16] - 2026-02-19
 
 ### Added